# Support normalized attributes