--- a/src/core/MeshBVH.js
+++ b/src/core/MeshBVH.js
@@ -189,7 +189,7 @@ export class MeshBVH {
 		const geometry = this.geometry;
 		const indexArr = geometry.index.array;
 		const posAttr = geometry.attributes.position;
-		
+
 		let buffer, uint32Array, uint16Array, float32Array;
 		let byteOffset = 0;
 		const roots = this._roots;
@@ -220,11 +220,13 @@ export class MeshBVH {
 				let maxx = - Infinity;
 				let maxy = - Infinity;
 				let maxz = - Infinity;
+
 				for ( let i = 3 * offset, l = 3 * ( offset + count ); i < l; i ++ ) {
+
 					const index = indexArr[ i ];
-					const x = posAttr.getX(index);
-					const y = posAttr.getY(index);
-					const z = posAttr.getZ(index);
+					const x = posAttr.getX( index );
+					const y = posAttr.getY( index );
+					const z = posAttr.getZ( index );
 
 					if ( x < minx ) minx = x;
 					if ( x > maxx ) maxx = x;