--- a/src/core/MeshBVH.js
+++ b/src/core/MeshBVH.js
@@ -189,6 +189,16 @@ export class MeshBVH {
 		const geometry = this.geometry;
 		const indexArr = geometry.index.array;
 		const posAttr = geometry.attributes.position;
+		const posArr = posAttr.array;
+
+		// support for an interleaved position buffer
+		const bufferOffset = posAttr.offset || 0;
+		let stride = 3;
+		if ( posAttr.isInterleavedBufferAttribute ) {
+
+			stride = posAttr.data.stride;
+
+		}
 
 		let buffer, uint32Array, uint16Array, float32Array;
 		let byteOffset = 0;
@@ -223,10 +233,25 @@ export class MeshBVH {
 
 				for ( let i = 3 * offset, l = 3 * ( offset + count ); i < l; i ++ ) {
 
-					const index = indexArr[ i ];
-					const x = posAttr.getX( index );
-					const y = posAttr.getY( index );
-					const z = posAttr.getZ( index );
+					let x;
+					let y;
+					let z;
+
+					if ( posAttr.normalized ) {
+
+						const index = indexArr[ i ];
+						x = posAttr.getX( index );
+						y = posAttr.getY( index );
+						z = posAttr.getZ( index );
+
+					} else {
+
+						const index = indexArr[ i ] * stride + bufferOffset;
+						x = posArr[ index + 0 ];
+						y = posArr[ index + 1 ];
+						z = posArr[ index + 2 ];
+
+					}
 
 					if ( x < minx ) minx = x;
 					if ( x > maxx ) maxx = x;