Initial pass at trying to add support for normalized attributes to `three-mesh-bvh`, to address #448. I only found a single instance in the code of an obvious, direct position attribute array access, so I changed it to use the setter. I'm assuming that `geometry.index` will never be normalized, so didn't touch any direct array accesses for that.

With this change, `shapecast` worked again with my small examples I'd created to debug this issue. I didn't yet test raycasting, or any other feature. All unit tests continue to pass.

The `MeshBVHVisualizer` is still broken, and doesn't display at all on meshes with normalized position attributes. I also still don't have as complete of an idea of how `three-mesh-bvh`'s internals work as I like, so any feedback or pointers are welcome. I also have done no performance testing.