@gkjohnson 2022-08-18
The position array is used directly when building the hierarchy, as well, so that will ahve to be changed: