@gkjohnson 2022-08-18
> With this change, shapecast worked again with my small examples I'd created to debug this issue.

Also are you using the latest dev version of three.js? These changes along will not do anything without the new fixes to three.