@dmliao 2022-08-18
Yes, in my own tests I was using the latest commit on the dev branch of three.js, which was how I saw that shapecast was working (at least visually? from playing around with it.)

And ah! I missed the hierarchy building part! I'll go and update that ASAP - I wonder if that's the reason why the visualizer is still broken?

Benchmarking build times is something I can do as well, once I update the hierarchy building. Are there any other places that directly reference the array?