@gkjohnson 2022-08-18
If the visualizer was broken then the hierarchy bounds were likely broken. So hopefully this fixes it.

> Are there any other places that directly reference the array?

I just searched for `.array` in the src folder and that looks like it for the most part? I don't think anything else requires it but maybe you can do a quick look to confirm. More eyes are always helpful!