@dmliao 2022-08-18
I've converted the array usage in `computeTriangleBounds` to the getters in the latest commit. This indeed fixes `MeshBVHVisualizer`.

Trying to compare build times didn't get me much...running `npm run build` is very fast on my machine (M1 Mac Pro), and the results were very close to each other:
```
# master branch
./src/index.js → ./build/index.umd.cjs...
created ./build/index.umd.cjs in 179ms

./src/index.js → ./build/index.module.js...
created ./build/index.module.js in 89ms

# this PR
./src/index.js → ./build/index.umd.cjs...

created ./build/index.umd.cjs in 167ms

./src/index.js → ./build/index.module.js...
created ./build/index.module.js in 88ms
```

I also ran `npm run benchmark`, here are the results:

[benchmark-original.txt](https://github.com/gkjohnson/three-mesh-bvh/files/9377872/benchmark-original.txt)
[benchmark-getters.txt](https://github.com/gkjohnson/three-mesh-bvh/files/9377876/benchmark-getters.txt)

It looks like using the getters does have a performance hit on most of the functionality.