@gkjohnson 2022-08-19
> It looks like using the getters does have a performance hit on most of the functionality, most obviously seen in the Compute BVH and DistanceToGeometry w/ BVH benchmarks.

`DistanceToGeometry` should not be affected since non of the code paths modified are related to that function. Only BVH building and the `refit` function should be affected. And in the benchmarks it looks like build time went form ~95ms to ~130ms (almost 30% increase) which is significant. Refit went up a good amount, as well.

Do you mind running the benchmarks 3 or so times each with only the meaningful tests and documenting the differences?

There's still some churn happening in three.js on this normalization change so we can wait until that settles but I propose testing retaining two code paths in the build and refit functions - one thats taken if the buffer is normalized and one thats taken if it is not. This should just involve an extra if statement in the for loop. Hopefully that won't impact the general performance quite as significantly.