@gkjohnson 2022-08-19
> Trying to compare build times didn't get me much...running npm run build is very fast on my machine (M1 Mac Pro), and the results were very close to each other:

Also to be clear these rollup / parcel build times will not be affected by the change. Only the runtime BVH construction times are important.