@dmliao 2022-08-24
After a slight delay, I ran a bunch of benchmarks against both versions, and averaged all of them, getting the following result:
```
(numbers in ms)

# NORMALIZED
            CENTER / AVERAGE / SAH
BVH         128 / 129 / 770    
Refit       10.7 / 10.6 / 10.9
Refit/Hint  0.4 / 2.7 / 0.3

# MASTER
BVH         95 / 94 / 736
Refit       9.3 / 9.2 / 9.5
Refit/Hint  0.4 / 2.4 / 0.3
```

Generally saw what I showed before, where the BVH construction time had a significant increase, and Refit a smaller time increase.

I haven't much time these past few days to dig into the code again, but agreed that given these performance numbers might be good to separate the code paths for normalized and unnormalized positions; it's slightly more complicated than just an if in the for loop, since several variables are set up differently ahead of time for them, but I'll figure it out.