@gkjohnson 2022-08-24
> agreed that given these performance numbers might be good to separate the code paths for normalized and unnormalized positions; it's slightly more complicated than just an if in the for loop, since several variables are set up differently ahead of time for them, but I'll figure it out.

Great thanks for the testing - and again I appreciate the time you're putting into this!